% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiLinearRegression.R
\name{MultiLinearRegression}
\alias{MultiLinearRegression}
\title{Multiple Linear regression function with Cholesky Decomposition}
\usage{
MultiLinearRegression(x, y, int = TRUE)
}
\arguments{
\item{x}{A size n x p Design matrix for regression (without intercept)}

\item{y}{A size n vector}

\item{int}{Does linear regression has intercept? TRUE for has intercept, FALSE for no intercept}
}
\value{
A list with the following elements:

beta.est :A vector of estimated beta for y ~ x *beta + error

Std.Error :A vector of standard error for each estimate

t.val :A vector of t-statistic for each estimate

p.value :A vector of p-values of each estimate

fitted :A vector of fitted values

SSE :A numeric value for the Sum of squared error of the model

R.square :A numeric value for the R square of the model
}
\description{
Multiple Linear regression function with Cholesky Decomposition
}
\examples{
x <- matrix(rnorm(1:300), 100, 3)
y <- rnorm(1:100)
fit = MultiLinearRegression(x,y,int = TRUE)
}
