% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnePassCov.R
\name{OnePassCov}
\alias{OnePassCov}
\title{One Pass Covariance between two numeric vectors}
\usage{
OnePassCov(x, y)
}
\arguments{
\item{x}{- A size n numeric vector}

\item{y}{- A size n numeric vector}
}
\value{
A numeric value of covariance between x and y with n-1 as denominator
}
\description{
One Pass Covariance between two numeric vectors
}
\examples{
x <- rnorm(1:100)
y <- rnorm(1:100)
OnePassCov(x,y)
}
