% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeRegression.R
\name{RidgeRegression}
\alias{RidgeRegression}
\title{The fast method to find beta estimates from Ridge Regression}
\usage{
RidgeRegression(X, Y, lambda, int = T)
}
\arguments{
\item{X}{A size n x p Design matrix for regression (without intercept)}

\item{Y}{A size n vector}

\item{lambda}{A numeric number, tuning parameter of the ridge regression model}

\item{int}{Does the regression has intercept? TRUE for has intercept, FALSE for no intercept}
}
\value{
A vector of estimated beta (to 4 precision) of the ridge regression model
}
\description{
The fast method to find beta estimates from Ridge Regression
}
\examples{
x <- matrix(rnorm(1:300), 100, 3)
y <- rnorm(1:100)
fit = RidgeRegression(x, y, lambda = 0.1, int = TRUE)
}
