% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SLM}
\alias{SLM}
\title{Simple Linear Regression function with Rcpp}
\arguments{
\item{x}{A size n predictor vector}

\item{y}{A size n outcome vector}
}
\value{
A list containing: intercept, beta estimate, std.Error, t.value, p.value from the simple linear regression
}
\description{
Simple Linear Regression function with Rcpp
}
